WIP: proptest finds the bug
Looks like we're not correctly handling a shadowing case in readdir if
the shadowing happens across a ListObjects request boundary (so we see
"a" in the first list and "a/" in the second). Not sure there's a ton we
can do in this case -- we already sent the bogus readdir reply by the
time we discover we messed up.